+++ src/bin/onetime-cli/main.rs
@@ -10,11 +10,17 @@ fn main() {
     match Subcommand::from(args) {
         Subcommand::Encrypt(e) => {
             onetime_cli::encrypt_file(&e).unwrap_or_exit(1);
-            println!("Successfully encrypted {}", e.file.to_string_lossy());
+
+            if !e.quiet {
+                println!("Successfully encrypted {}", e.file.to_string_lossy());
+            }
         }
         Subcommand::Decrypt(d) => {
             onetime_cli::decrypt_file(&d).unwrap_or_exit(1);
-            println!("Successfully decrypted {}", d.file.to_string_lossy());
+
+            if !d.quiet {
+                println!("Successfully decrypted {}", d.file.to_string_lossy());
+            }
         }
         Subcommand::None => {
             let mut cmd = clap_app::build_clap_app();